+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ENHANCEMENTS:
 FEATURES:
 * compute: added support for `instance_tags_pool` in `yandex_compute_instance_group`
 * compute: added support for `ignore_health_checks` in `yandex_compute_instance_group`
+* managed-kubernetes: add `master_location`, `etcd_cluster_size` in `yandex_kubernetes_cluster` resource and data source
 * kafka: support Kafka 3.3,3.4,3.5 in `yandex_mdb_kafka_cluster`
 
 * **New Resource:** `yandex_loadtesting_agent`